Add qrc20 tests to docker_tests * Replace Qtum env bootstrap into separated qrc20_tests module
